Moved settings models
