+++ Mamesaver/Configuration/Models/FontSettings.cs
@@ -1,0 +1,15 @@
+namespace Mamesaver.Configuration.Models
+{
+    public class FontSettings
+    {
+        /// <summary>
+        ///     Font face.
+        /// </summary>
+        public string Face { get; set; }
+
+        /// <summary>
+        ///     Font size, in points.
+        /// </summary>
+        public int Size { get; set; }
+    }
+}
