--- Mamesaver/Models/Settings/FontSettings.cs
@@ -1,8 +1,0 @@
-namespace Mamesaver.Models.Settings
-{
-    public class FontSettings
-    {
-        public string Face { get; set; }
-        public int Size { get; set; }
-    }
-}
